+++ src/views/TechnicalAssistance/RequestForm/Documents.tsx
@@ -1,4 +1,4 @@
-import React, { useEffect, useMemo, useState } from 'react';
+import React, { useEffect, useState } from 'react';
 import { Controller, useForm } from 'react-hook-form';
 import { useTranslation } from 'react-i18next';
 import {
@@ -8,8 +8,7 @@ import {
   useParams,
   useRouteMatch
 } from 'react-router-dom';
-import { CellProps, Column, useSortBy, useTable } from 'react-table';
-import { useMutation, useQuery } from '@apollo/client';
+import { ApolloQueryResult, useMutation } from '@apollo/client';
 import { yupResolver } from '@hookform/resolvers/yup';
 import {
   Alert,
@@ -22,38 +21,23 @@ import {
   Grid,
   IconArrowBack,
   Label,
-  Link,
   Radio,
-  Table,
   TextInput
 } from '@trussworks/react-uswds';
 import { clone } from 'lodash';
-import { DateTime } from 'luxon';
 
 import UswdsReactLink from 'components/LinkWrapper';
 import PageHeading from 'components/PageHeading';
 import Spinner from 'components/Spinner';
 import CreateTrbRequestDocumentQuery from 'queries/CreateTrbRequestDocumentQuery';
-import DeleteTrbRequestDocumentQuery from 'queries/DeleteTrbRequestDocumentQuery';
-import GetTrbRequestDocumentsQuery from 'queries/GetTrbRequestDocumentsQuery';
 import {
   CreateTrbRequestDocument,
   CreateTrbRequestDocumentVariables
 } from 'queries/types/CreateTrbRequestDocument';
 import {
-  DeleteTrbRequestDocument,
-  DeleteTrbRequestDocumentVariables
-} from 'queries/types/DeleteTrbRequestDocument';
-import {
   GetTrbRequestDocuments,
-  GetTrbRequestDocuments_trbRequest_documents as TrbRequestDocuments,
   GetTrbRequestDocumentsVariables
 } from 'queries/types/GetTrbRequestDocuments';
-import {
-  TRBDocumentCommonType,
-  TRBRequestDocumentStatus
-} from 'types/graphql-global-types';
-import { getColumnSortStatus, getHeaderSortIcon } from 'utils/tableSort';
 import {
   documentSchema,
   TrbRequestInputDocument
@@ -61,8 +45,13 @@ import {
 
 import Breadcrumbs from '../Breadcrumbs';
 
+import DocumentsTable from './DocumentsTable';
 import Pager from './Pager';
 import { FormStepComponentProps, StepSubmit } from '.';
+
+export type RefetchDocuments = (
+  variables?: Partial<GetTrbRequestDocumentsVariables> | undefined
+) => Promise<ApolloQueryResult<GetTrbRequestDocuments>> | (() => void);
 
 /**
  * Documents is a component of both the table list of uploaded documents
@@ -77,21 +66,13 @@ function Documents({
 }: FormStepComponentProps) {
   const { t } = useTranslation('technicalAssistance');
   const { t: gt } = useTranslation('general');
+
   const history = useHistory();
   const { url } = useRouteMatch();
 
   const { view } = useParams<{
     view?: string;
   }>();
-
-  const { data, refetch, loading } = useQuery<
-    GetTrbRequestDocuments,
-    GetTrbRequestDocumentsVariables
-  >(GetTrbRequestDocumentsQuery, {
-    variables: { id: request.id }
-  });
-
-  const documents = data?.trbRequest.documents || [];
 
   // Documents can be created from the upload form
   const [createDocument] = useMutation<
@@ -100,108 +81,6 @@ function Documents({
   >(CreateTrbRequestDocumentQuery);
 
   const [isUploadError, setIsUploadError] = useState(false);
-
-  // Documents can be deleted from the table
-  const [deleteDocument] = useMutation<
-    DeleteTrbRequestDocument,
-    DeleteTrbRequestDocumentVariables
-  >(DeleteTrbRequestDocumentQuery);
-
-  const columns = useMemo<Column<TrbRequestDocuments>[]>(() => {
-    return [
-      {
-        Header: t<string>('documents.table.header.fileName'),
-        accessor: 'fileName'
-      },
-      {
-        Header: t<string>('documents.table.header.documentType'),
-        accessor: ({ documentType: { commonType, otherTypeDescription } }) => {
-          if (commonType === TRBDocumentCommonType.OTHER) {
-            return otherTypeDescription || '';
-          }
-          return t(`documents.upload.type.${commonType}`);
-        }
-      },
-      {
-        Header: t<string>('documents.table.header.uploadDate'),
-        accessor: 'uploadedAt',
-        Cell: ({ value }) => DateTime.fromISO(value).toFormat('MM/dd/yyyy')
-      },
-      {
-        Header: t<string>('documents.table.header.actions'),
-        accessor: ({ status }) => {
-          // Repurpose the accessor to use `status` for sorting order
-          if (status === TRBRequestDocumentStatus.PENDING) return 1;
-          if (status === TRBRequestDocumentStatus.AVAILABLE) return 2;
-          if (status === TRBRequestDocumentStatus.UNAVAILABLE) return 3;
-          return 4;
-        },
-        Cell: ({ row }: CellProps<TrbRequestDocuments, string>) => {
-          // Show the upload status
-          // Virus scanning
-          if (row.original.status === TRBRequestDocumentStatus.PENDING)
-            return (
-              <em data-testurl={row.original.url}>
-                {t('documents.table.virusScan')}
-              </em>
-            );
-          // View or Remove
-          if (row.original.status === TRBRequestDocumentStatus.AVAILABLE)
-            // Show some file actions once it's available
-            return (
-              <>
-                {/* View document */}
-                <Link target="_blank" href={row.original.url}>
-                  {t('documents.table.view')}
-                </Link>
-                {/* Delete document */}
-                <Button
-                  unstyled
-                  type="button"
-                  className="margin-left-2 text-error"
-                  onClick={() => {
-                    deleteDocument({
-                      variables: { id: row.original.id }
-                    })
-                      .then(() => {
-                        refetch(); // Refresh doc list
-                      })
-                      .catch(() => {
-                        // todo no top level error message for the delete yet
-                      });
-                  }}
-                >
-                  {t('documents.table.remove')}
-                </Button>
-              </>
-            );
-          // Infected unavailable
-          if (row.original.status === TRBRequestDocumentStatus.UNAVAILABLE)
-            return t('documents.table.unavailable');
-          return '';
-        }
-      }
-    ];
-  }, [deleteDocument, refetch, t]);
-
-  const {
-    getTableBodyProps,
-    getTableProps,
-    headerGroups,
-    prepareRow,
-    rows
-  } = useTable(
-    {
-      columns,
-      data: documents,
-      autoResetSortBy: false,
-      autoResetPage: false,
-      initialState: {
-        sortBy: useMemo(() => [{ id: 'uploadedAt', desc: true }], [])
-      }
-    },
-    useSortBy
-  );
 
   const {
     control,
@@ -217,6 +96,12 @@ function Documents({
       otherTypeDescription: ''
     }
   });
+
+  const [documentsCount, setDocumentsCount] = useState(0);
+  const [refetchDocuments, setRefetchDocuments] = useState<RefetchDocuments>(
+    () => () => {}
+  );
+  const [loadingDocuments, setLoadingDocuments] = useState(true);
 
   const submit = handleSubmit(async formData => {
     const input: any = clone(formData);
@@ -235,7 +120,7 @@ function Documents({
       }
     })
       .then(() => {
-        refetch(); // Reload documents
+        refetchDocuments(); // Reload documents
         setFormAlert({
           type: 'success',
           slim: true,
@@ -245,6 +130,7 @@ function Documents({
         history.push(`/trb/requests/${request.id}/documents`);
       })
       .catch(err => {
+        // console.log(err);
         setIsUploadError(true);
       });
   });
@@ -284,103 +170,54 @@ function Documents({
     <Switch>
       {/* Documents table */}
       <Route exact path="/trb/requests/:id/documents">
-        {loading ? (
+        {loadingDocuments && (
           <div className="margin-y-10" data-testid="page-loading">
             <div className="text-center">
               <Spinner size="xl" aria-valuetext={gt('pageLoading')} aria-busy />
             </div>
           </div>
-        ) : (
-          <>
-            {/* Open the document upload form */}
-            <div className="margin-top-5 margin-bottom-4">
-              <UswdsReactLink
-                variant="unstyled"
-                className="usa-button"
-                to={`${url}/upload`}
-              >
-                {t('documents.addDocument')}
-              </UswdsReactLink>
-            </div>
+        )}
 
-            <div>
-              <Table bordered={false} fullWidth scrollable {...getTableProps()}>
-                <thead>
-                  {headerGroups.map(headerGroup => (
-                    <tr {...headerGroup.getHeaderGroupProps()}>
-                      {headerGroup.headers.map((column, index) => (
-                        <th
-                          {...column.getHeaderProps(
-                            column.getSortByToggleProps()
-                          )}
-                          aria-sort={getColumnSortStatus(column)}
-                          scope="col"
-                          className="border-bottom-2px"
-                        >
-                          <Button
-                            type="button"
-                            unstyled
-                            className="width-full display-flex"
-                            {...column.getSortByToggleProps()}
-                          >
-                            <div className="flex-fill text-no-wrap">
-                              {column.render('Header')}
-                            </div>
-                            <div className="position-relative width-205 margin-left-05">
-                              {getHeaderSortIcon(column)}
-                            </div>
-                          </Button>
-                        </th>
-                      ))}
-                    </tr>
-                  ))}
-                </thead>
-                <tbody {...getTableBodyProps()}>
-                  {rows.map(row => {
-                    prepareRow(row);
-                    return (
-                      <tr {...row.getRowProps()}>
-                        {row.cells.map((cell, index) => {
-                          return (
-                            <td {...cell.getCellProps()}>
-                              {cell.render('Cell')}
-                            </td>
-                          );
-                        })}
-                      </tr>
-                    );
-                  })}
-                </tbody>
-              </Table>
-              {data && documents.length === 0 && (
-                <div className="font-body-2xs margin-left-2">
-                  {t('documents.table.noDocument')}
-                </div>
-              )}
-            </div>
+        {/* Open the document upload form */}
+        <div className="margin-top-5 margin-bottom-4">
+          <UswdsReactLink
+            variant="unstyled"
+            className="usa-button"
+            to={`${url}/upload`}
+          >
+            {t('documents.addDocument')}
+          </UswdsReactLink>
+        </div>
 
-            <Pager
-              className="margin-top-7"
-              back={{
-                onClick: () => {
-                  history.push(stepUrl.back);
-                }
-              }}
-              next={{
-                onClick: e => {
-                  history.push(stepUrl.next);
-                },
-                text: t(
-                  documents.length
-                    ? 'button.next'
-                    : 'documents.continueWithoutAdding'
-                ),
-                outline: documents.length === 0
-              }}
-              submit={submitNoop}
-              taskListUrl={taskListUrl}
-            />
-          </>
+        <DocumentsTable
+          trbRequestId={request.id}
+          setLoadingDocuments={setLoadingDocuments}
+          setRefetchDocuments={setRefetchDocuments}
+          setDocumentsCount={setDocumentsCount}
+        />
+
+        {!loadingDocuments && (
+          <Pager
+            className="margin-top-7"
+            back={{
+              onClick: () => {
+                history.push(stepUrl.back);
+              }
+            }}
+            next={{
+              onClick: e => {
+                history.push(stepUrl.next);
+              },
+              text: t(
+                documentsCount
+                  ? 'button.next'
+                  : 'documents.continueWithoutAdding'
+              ),
+              outline: documentsCount === 0
+            }}
+            submit={submitNoop}
+            taskListUrl={taskListUrl}
+          />
         )}
       </Route>
